fontend: Move confetti easter egg to its own module
Move confetti easter egg to its own module - this is only a code moving
commit, we will clean-up the code in a follow-up commit.
