+++ frontend/index.js
@@ -1,5 +1,6 @@
 "use strict"
 import Editor from "./module/editor.js"
+import showConfetti from "./module/confetti.js"
 
 // --- Globals ---------------------------------------------------------
 
@@ -119,6 +120,7 @@ function jsPrint(ptr, len) {
   const output = document.querySelector("#console")
   output.textContent += s
   output.scrollTo({ behavior: "smooth", left: 0, top: output.scrollHeight })
+  // 🐣 Show confetti Easter egg if print argument contains literal string "confetti"
   if (s.toLowerCase().includes("confetti")) {
     showConfetti()
   }
@@ -952,61 +954,6 @@ function showAbout() {
   const about = document.querySelector("#dialog-about")
   hideSidemenu()
   about.showModal()
-}
-
-// --- UI: Confetti Easter Egg -----------------------------------------
-//
-// When code input string contains the sub string "confetti" show
-// confetti on Run button click.
-
-function showConfetti() {
-  const names = ["🦊", "🐐"]
-  const colors = ["red", "purple", "blue", "orange", "gold", "green"]
-  let confetti = new Array(100)
-    .fill()
-    .map((_, i) => {
-      return {
-        name: names[i % names.length],
-        x: Math.random() * 100,
-        y: -20 - Math.random() * 100,
-        r: 0.1 + Math.random() * 1,
-        color: colors[i % colors.length],
-      }
-    })
-    .sort((a, b) => a.r - b.r)
-
-  const cssText = (c) =>
-    `background: ${c.color}; left: ${c.x}%; top: ${c.y}%; transform: scale(${c.r})`
-  const confettiDivs = confetti.map((c) => {
-    const div = document.createElement("div")
-    div.style.cssText = cssText(c)
-    div.classList.add("confetti")
-    div.textContent = c.name
-    document.body.appendChild(div)
-    return div
-  })
-
-  let frame
-
-  function loop() {
-    frame = requestAnimationFrame(loop)
-    confetti = confetti.map((c, i) => {
-      c.y += 0.7 * c.r
-      if (c.y > 120) c.y = -20
-      const div = confettiDivs[i]
-      div.style.cssText = cssText(c)
-      return c
-    })
-  }
-
-  loop()
-  setTimeout(() => {
-    cancelAnimationFrame(frame)
-    confettiDivs.forEach((div) => div.remove())
-  }, 10000)
-  setTimeout(() => {
-    confettiDivs.forEach((div) => div.classList.add("fadeout"))
-  }, 8500)
 }
 
 // --- Share / load snippets -------------------------------------------
